+++ TestSutraUsers/src/main/java/com/tathvatech/project/oid/ProjectStageOID.java
@@ -2,10 +2,11 @@ package com.tathvatech.project.oid;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.tathvatech.ts.core.common.EntityTypeEnum;
-import com.tathvatech.ts.core.common.OID;
 
-public class ProjectStageOID extends OID{
+import com.tathvatech.common.enums.EntityTypeEnum;
+import com.tathvatech.user.OID.OID;
+
+public class ProjectStageOID extends OID {
 
 	public ProjectStageOID()
 	{
@@ -25,7 +26,7 @@ public class ProjectStageOID extends OID{
 
 	@Override
 	public int hashCode() {
-		return getPk();
+		return (int) getPk();
 	}
 
 	@Override
